[FEAT] Enhance API documentation and add image processing functions
- Updated API documentation to reflect changes in image processing modules.
- Added functions for parcellating, registering, and transforming images in the images.py file.
- Introduced new io.py and regions.py files with functions for loading and remapping data and querying structure graphs, respectively.
- Removed obsolete transforms.py file.
