+++ neuromaps_mouse/stats.py
@@ -2,4 +2,10 @@
 
 
 def correlation():
+    """Compute spatial correlation between two brain maps."""
+    pass
+
+
+def moran():
+    """Compute Moran's I spatial autocorrelation."""
     pass
