--- neuromaps_mouse/transforms.py
@@ -1,4 +1,0 @@
-
-
-def allenccfv3_to_allenccfv3():
-    pass
